requirements.txt
apache-airflow==2.9.0
dbt-core==1.10.4
dbt-snowflake==1.10.0
